fix cache issue
