+++ src/Conversion/ResetCache.php
@@ -16,21 +16,12 @@ class ResetCache
 
             $currencies = CurrencyRate::query()->pluck('currency');
             foreach ($currencies as $currency) {
-                Cache::forget(self::cacheName($currency));
+                Cache::forget(Convert::cacheName($currency));
             }
             return "The converter cache has been reset successfully!";
         } catch (\Exception $exception) {
             report($exception);
             return "Failed this command";
         }
-    }
-
-    /**
-     * @param string $currency
-     * @return string
-     */
-    private static function cacheName(string $currency): string
-    {
-        return ConvertRate::cacheName($currency);
     }
 }
